working on pro pre ver

+++ app/src/main/java/com/teamx/hatlyUser/ui/fragments/products/model/Veriation.kt
@@ -4,7 +4,7 @@ import androidx.annotation.Keep
 @Keep
 data class Veriation(
     val options: List<Option>,
-    val prize: Int,
+    val prize: Double,
     val quantity: Int,
     val title: String
 )
